refactor: Update import paths for LoginError and LoginUiState to improve module organization
